# userservices/.optic/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Optic Generated OpenAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/lender/getDashboardForLender": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "TotalinvoicesCount": {
                          "type": "number"
                        },
                        "TotalfundedPaidAmount": {
                          "type": "number"
                        },
                        "TotalfundedCount": {
                          "type": "number"
                        },
                        "TotalRepaymentsAmount": {
                          "type": "number"
                        },
                        "TotalInvoicesAmount": {
                          "type": "number"
                        },
                        "TotalHospital": {
                          "type": "number"
                        },
                        "Repaymentcount": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "success",
                        "TotalinvoicesCount",
                        "TotalfundedPaidAmount",
                        "TotalfundedCount",
                        "TotalRepaymentsAmount",
                        "TotalInvoicesAmount",
                        "TotalHospital",
                        "Repaymentcount"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lender/getLenderGraphOne": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "monthsAndYears": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          }
                        },
                        "Repaid": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "Rejected": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "InProcesss": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "Disbursed": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "All": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        }
                      },
                      "required": [
                        "success",
                        "monthsAndYears",
                        "Repaid",
                        "Rejected",
                        "InProcesss",
                        "Disbursed",
                        "All"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lender/getInvoiceByIdToLender": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "nullable": true
                        }
                      },
                      "required": [
                        "success",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/lender/getAllInvoicesToLender": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "numberOfPages": {
                          "type": "number"
                        },
                        "invoices": {
                          "type": "array",
                          "items": {}
                        }
                      },
                      "required": [
                        "numberOfPages",
                        "invoices"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/lender/getHospitalToLender": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "numberOfPages": {
                          "type": "number"
                        },
                        "message": {
                          "type": "array",
                          "items": {}
                        }
                      },
                      "required": [
                        "success",
                        "numberOfPages",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lender/updatePassword": {
      "put": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "success"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/lender/getTransactionDataforLender": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "array",
                          "items": {}
                        }
                      },
                      "required": [
                        "success",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lender/postFromLender": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string"
                  },
                  "PFFees": {
                    "type": "number"
                  },
                  "LenderComment": {
                    "type": "string"
                  },
                  "Interest": {
                    "type": "number"
                  },
                  "ApporvedAmount": {
                    "type": "number"
                  },
                  "AmountToBePaid": {
                    "type": "number"
                  }
                },
                "required": [
                  "_id",
                  "PFFees",
                  "LenderComment",
                  "Interest",
                  "ApporvedAmount",
                  "AmountToBePaid"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "success",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}